refactor(neuron-wallet): reorganize neuron-wallet files and add fullscreen on the mainmenu
