+++ packages/neuron-wallet/src/services/cells.ts
@@ -1,5 +1,4 @@
 import { getConnection, In } from 'typeorm'
-import asw from '../wallets/asw'
 import { getLiveCells } from '../mock_rpc'
 import OutputEntity from '../entities/Output'
 import { Input } from './transactions'
@@ -167,10 +166,5 @@ export default class CellsService {
       inputs,
       capacities: inputCapacities.toString(),
     }
-  }
-
-  public static getUnspentCells = async () => {
-    const cells = await asw.getUnspentCells()
-    return cells
   }
 }
