--- packages/neuron-wallet/src/wallets/asw.ts
@@ -1,3 +1,0 @@
-import ckbCore from '../core'
-
-export default ckbCore.wallet.newASW()
